support for the options --warn-undefined-functions
--warn-undefined-variables do warn in case of attempt to call an
undefined user function, but there can be a lot of legitimate
undefined variables, so the useful warnings get lost in the crowd

on the other hand calling an undefined user function is rarely legitimate
and can easily be silenced.

this patch add the support for a new options --warn-undefined-functions
that produce a warning when $(call is invoked on an undefined variable.