script to create git tree with submodule setup from bundles
the existing script doesn't do anything useful for current development
line, it is meant for the old buildsetup where there was a clone
directory and tons of symlinks.

Unfortunately using git bundles with submodules is not just unpacking
the tarballs, the git config needs to be tweaked for that. Hence this
script that downloads the bundles and does all the work.

./lo-submodules-from-bundles.sh -t my-checkout-name help dict

to setup the checkout with the helpcontent2 and dictionaries submodules.
For more details, see ./lo-submodules-from-bundles.sh -h

nice benefit over using git is that the script setups the repositories
using relative paths, so you can move/copy your checkout without git
complaining

Change-Id: Ia483039fcbafe8fc042877d6b610ee88a6097c0a
Reviewed-on: https://gerrit.libreoffice.org/3653
Reviewed-by: Thorsten Behrens <tbehrens@suse.com>
Tested-by: Thorsten Behrens <tbehrens@suse.com>