Add NSAXSpy to help debugging OS X accessibility
This tool currently logs the AXAttributedString of the first AXTextArea
found in LibreOffice (suited for Writer), and logs all AXValueChanged
and AXSelectedTextChanged notifications of that AXTextArea and also all
AXFocusedUIElementChange notifications. It can be adjusted to do
the same thing for TextEdit so that one can learn from it
 how stuff should be implemented (TextEdit is basically
*the* reference implementation of AXTextArea).

Any adjustments (e.g. to observe TextEdit instead of LibreOffice)
need to be done in the source code of this tool.

Change-Id: Ia7f4dbed4924b8f5330726d378eda3601991f404
Reviewed-on: https://gerrit.libreoffice.org/5160
Reviewed-by: Boris Dušek <me@dusek.me>
Tested-by: Boris Dušek <me@dusek.me>