formatxml: avoid corrupting ODF paragraphs / headings
The problem was that

<text:p>a<text:span>b</text:span>c</text:p>

was formatted as

      <text:p text:style-name="P1">He
        <text:span text:style-name="T1">ll</text:span>
        o</text:p>

So the resulting imported paragraph string changed from "Hello" to "He ll o".

If there is situation where formatting children of text:p / text:h is wanted,
it would be possible to add a cmdline option to disable this behavior.

Also, for now this assumes that all children of these elements should
not be formatted to be on the safe side.

Change-Id: Ifcb846eef594d53206fa870b5eb31afabba7b5ae