tdf#94167 Modifying fields turm them into Expression is faulty
The root cause of the error is, that xmlhelp2soffice.xsl has produced a wrong
nesting of office:text and text:variable-decls. That has caused, that the
field edit dialog does not know the format of the field and therefore does
not accept text format. The patch corrects the nesting.

Change-Id: I90a1cfee41a503a4a75e7b71d7e4688ae58847d5
Reviewed-on: https://gerrit.libreoffice.org/18766
Reviewed-by: Adolfo Jayme Barrientos <fitojb@ubuntu.com>
Tested-by: Adolfo Jayme Barrientos <fitojb@ubuntu.com>