tdf#93981 Attribute localize=(false|true) is deleted
The information localize='false' is tracked by adding _NOL10N to
the paragraph id. But some files exist, where the paragraph does not
have an id at all. In such cases an id is created, as an id is
a required attribute. This solves tdf#94612 partly, because a
missing id in a file would otherwise generate an empty id on import,
which is neither allowed.

Change-Id: Ib136493dee004fd09f8e7e14d05c99ea173865e0
Reviewed-on: https://gerrit.libreoffice.org/19166
Reviewed-by: Adolfo Jayme Barrientos <fitojb@ubuntu.com>
Tested-by: Adolfo Jayme Barrientos <fitojb@ubuntu.com>