New online help content
Notice:
1) The folder text/ is exactly the same as in helpcontent2/ minus the first line.
The first line has the <xml> declaration followed by the <xsl> declaration. NOTHING MORE.

2) the online_transform.xsl is main_transform.xsl of HC2 modified to accomodate the
layout of the screen, It has 4 <div> hooked in each of the 4 corners of the screen to
allow resizing.

3) the lower-right <div> displays the help content
4) the upper-left <div> has entries to change help books/modules
5) the upper right <div> is available to place a search mechanism
6) the lower-left <div> is available to inser a index with javascript

7) index.html is just a top landing page.

8) default.css is the css used in helpconten2/languages

9) plenty of simplification/tweak is possible in the XSLT.

10) The whole HC can be tranformed in html with XSLTPROC or any other XSL
processor, provided the reference to the xsl is fixed, and the xsl file be aware of the
specifics of the xsl processor.

11) Help pages can be acessed by http://<url>/help_file.xhp#<bookmark | section | whatever ID>

12) Works with Firefox, Chrome, Rekonq. MS Edge does not understand xml+xsl at first glance,
likely to need some config, I guess.

13) References to media are incomplete. That will require a change in the urls of the
media inside the help pages, thus forking from HC2