Explicitly list 3rd-party dependencies in the manifest file
...instead of relying on 'make fetch' during the build (which allows to drop
the --share=network build-options override).  The list is based on the work
done at ramcq/libreoffice-flatpak on github.com, but adapted to current LO
master.  The plan is to ultimately move the manifest into the core LO git repo
and share it between this flatpak/build.sh and Flathub builds (for which not
downloading stuff during the build itself is a prerequisite).