xhp-editor, further work
Hello,

I've had some more time to work on the editor -- the latest version is
attached! My changes:

* Completed the XHP syntax (see autocomplete.js -- and note that
entities and attributes with dashes need to be surrounded with quotes)

* Fixed the problem with "meta" not closing properly, as you saw in the
video a while back (for future reference, the problem was caused in
mode/xml/xml.js: line 15: "autoSelfClosers", which had "meta" defined.
If any other XHP entities have this problem, check there first)

* Removed various unneeded CodeMirror files, to simplify the structure
and hopefully make it easier for others to get involved

* Some cosmetic cleanups

Change-Id: If536db1e1c6dd13a6088d72658b79843d32edc44
Reviewed-on: https://gerrit.libreoffice.org/60034
Reviewed-by: Olivier Hallot <olivier.hallot@libreoffice.org>
Tested-by: Olivier Hallot <olivier.hallot@libreoffice.org>