list-ci-changes: add cmdline tool to list not yet queued gerrit changes
If you push a change to gerrit and the build queue is long enough, then
there is no "Build queued" on it, but it's not ignored, either.

This tool can list the changes which are in this "not yet queued" state,
without having to interact with the hard-to-understand
ci.libreoffice.org in a browser.

Change-Id: Ibe7579d911741dce491d5538ab76639d453769dc