ciabot: store repositories into a dedicated directory
(Hardcoded to ‘/path/to/../repositories’, where ‘/path/to/run-libreoffice-ciabot.pl’
is the full path to the daemon.)

This is useful to harden the the service file: ciabot needs write access
to repositories, but neither to its homedir nor to its codebase.

Other scripts such as libreoffice-bugzilla2.py are now expected to be
found under ‘/path/to’ as well.  Previously they were looked up in the
current directory.