git-cherry-gerrit.py: add a python version of the bash script
A functional change from the bash version is that we stop after the
first Change-Id when reading a commit message. This should not cause too
much problems (only happens when directly pushing multiple commits to
e.g. help.git and gerrit generating a submodule update for core.git),
and simplifies the code.

The benefit is that it's easy to use 'git cat-file --batch' from Python,
which gives a performance boost:

28,140 vs 6,105 seconds when comparing e.g. distro/collabora/cp-6.4 and
master of core.git (roughly 5 times faster).

Change-Id: I9ec91ed28bfdbc843f53a9e9f1a7ddd853e63082
Reviewed-on: https://gerrit.libreoffice.org/c/dev-tools/+/109077
Tested-by: Miklos Vajna <vmiklos@collabora.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>